sender1/requirement.txt
adal==1.2.2
alabaster==0.7.12
amqp==2.5.1
appdirs==1.4.3
argon2-cffi==19.1.0
aspy.yaml==1.3.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
azure-common==1.1.24
azure-servicebus==0.50.2
Babel==2.7.0
backcall==0.1.0
billiard==3.6.1.0
black==19.3b0
celery==4.3.0
certifi==2019.9.11
cffi==1.13.2
cfgv==2.0.1
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
cryptography==2.8
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.4
django-allauth==0.39.1
django-appconf==1.0.3
django-autocomplete-light==3.4.1
django-celery-beat==1.5.0
django-ckeditor==5.7.1
django-compressor==2.3
django-cors-headers==2.0.2
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.2.1
django-filter==2.2.0
django-import-export==1.2.0
django-js-asset==1.2.2
django-jsonfield==1.4.0
django-model-utils==3.2.0
django-notifications-hq==1.5.0
django-notify-x==0.1.9
django-redis==4.10.0
django-timezone-field==3.1
django-webpack-loader==0.6.0
djangorestframework==3.10.2
docutils==0.15.2
drf-yasg==1.17.0
entrypoints==0.3
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==2.0.3
flake8==3.7.8
flower==0.9.3
identify==1.4.7
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
importlib-resources==1.0.2
inflection==0.3.1
ipdb==0.12.2
ipython==7.9.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.21
itypes==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
jsonfield==2.0.2
kombu==4.6.5
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
msrest==0.6.10
msrestazure==0.6.2
mypy==0.720
mypy-extensions==0.4.3
nodeenv==1.3.3
numpy==1.17.4
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.0
packaging==19.2
parso==0.5.1
pathspec==0.6.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.13.0
pre-commit==1.20.0
prompt-toolkit==2.0.10
psycopg2==2.8.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.4.3
pylint-celery==0.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
pyparsing==2.4.4
pytest==5.0.1
pytest-cov==2.8.1
pytest-django==3.5.1
pytest-sugar==0.9.2
python-crontab==2.4.0
python-dateutil==2.8.1
python-slugify==3.0.3
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
rcssmin==1.0.6
redis==3.3.6
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
tablib==0.14.0
termcolor==1.1.0
text-unidecode==1.2
toml==0.10.0
tornado==5.1.1
traitlets==4.3.3
typed-ast==1.4.0
typing-extensions==3.7.4.1
uamqp==1.2.5
uritemplate==3.0.0
urllib3==1.25.6
vine==1.3.0
virtualenv==16.7.7
wcwidth==0.1.7
Werkzeug==0.14.1
whitenoise==4.1.3
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
